refactor: VideoPlaybackBar 컴포넌트 분리 (#47)
- ProgressBar.tsx: 재생 프로그레스바 + 스크러빙 + 썸네일 미리보기
- VolumeControl.tsx: 볼륨 조절 + 시간 표시
- VideoPlaybackBar: 조합 컴포넌트로 역할 축소 (327줄 → 152줄)
